(Hopefully) fixed a crash, a deadlock, and a memory leak all in one
Crash: some video files have negative audio PTSes at the very beginning.
Our code previously couldn't handle that, now it can, as long as the
total timestamp is positive.  TODO: figure out how to remove that
restriction.

Deadlock: When a cpal::Stream is dropped, it join()s the cpal thread.
Previously, we blocked the cpal thread waiting for audio packets from
the decoder thread.  We no longer do; instead, we fill with silence if
an audio frame is not immediately available. This seems to have resolved
the issue.

Memory leak: previously, video_decode_thread() could never return,
meaning that the Input struct could never be dropped.  I added Drop code
to VideoPlayerWidget that calls close_read(), which *should* stop writes
from happening.  With the current implementation, that should cause the
decode thread to panic when the VideoPlayerWidget gets destroyed, which
is its own problem.  More importantly, that panic doesn't seem to be
happening, so I'm not sure it worked.  Still a lot to do.